pyproject.toml
[project]
name = "sgcd"
version = "0.1.0"
description = ""
authors = [
    {name = "carloscasimirocc",email = "carloscasimirocc@yahoo.com.br"}
]
readme = "README.md"
requires-python = ">=3.12, <4"
dependencies = [
    "alembic (==1.16.4)",
    "annotated-types (==0.7.0)",
    "astroid (==3.3.11)",
    "autoflake (==2.3.1)",
    "autopep8 (==2.3.2)",
    "black (==25.1.0)",
    "blinker (==1.9.0)",
    "blueprint (==3.4.2)",
    "certifi (==2025.10.5)",
    "charset-normalizer (==3.4.4)",
    "click (==8.2.1)",
    "colorama (==0.4.6)",
    "dill (==0.4.0)",
    "dnspython (==2.8.0)",
    "ecdsa (==0.19.1)",
    "email-validator (==2.3.0)",
    "flake8 (==7.3.0)",
    "flask (==3.1.1)",
    "flask-admin (==1.6.1)",
    "flask-jwt-extended (==4.7.1)",
    "flask-login (==0.6.3)",
    "flask-mailman (==1.1.1)",
    "flask-marshmallow (==1.3.0)",
    "flask-migrate (==4.1.0)",
    "flask-sqlalchemy (==3.1.1)",
    "flask-wtf (==1.2.2)",
    "greenlet (==3.2.3)",
    "idna (==3.10)",
    "isort (==6.0.1)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.6)",
    "mako (==1.3.10)",
    "markupsafe (==3.0.2)",
    "marshmallow (==4.0.1)",
    "marshmallow-sqlalchemy (==1.0.0)",
    "mccabe (==0.7.0)",
    "mypy-extensions (==1.1.0)",
    "numpy (==2.3.2)",
    "packaging (==25.0)",
    "pathspec (==0.12.1)",
    "pillow (==12.0.0)",
    "platformdirs (==4.3.8)",
    "psycopg2-binary (==2.9.10)",
    "pyarrow (==21.0.0)",
    "pycodestyle (==2.14.0)",
    "pydantic (==2.12.4)",
    "pydantic-core (==2.41.5)",
    "pyflakes (==3.4.0)",
    "pyjwt (==2.10.1)",
    "pylance (==0.32.0)",
    "pylint (==3.3.7)",
    "python-dotenv (==1.1.1)",
    "python-http-client (==3.3.7)",
    "reportlab (==4.4.4)",
    "requests (==2.32.5)",
    "six (==1.17.0)",
    "sqlalchemy (==2.0.42)",
    "tomlkit (==0.13.3)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.14.1)",
    "urllib3 (==2.5.0)",
    "werkzeug (==3.1.3)",
    "wtforms (==3.2.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
